+++ src/result-validators/fixable.ts
@@ -1,0 +1,11 @@
+import { nameOf, objectHasFunctions } from "../utils";
+
+export interface FixableButton {
+    id: string;
+    label: string;
+}
+
+export interface Fixable {
+    getFixButtons(): Array<FixableButton>;
+    startFix(buttonId: string): void;
+}